Added trip backend Moved pinpoint to own controller Made password 100 characters instead of 20. This is for more safety
