+++ Project-Starlord/Helpers/ExtensionMethods.cs
@@ -26,7 +26,7 @@ namespace Project_Starlord.Helpers
             new RNGCryptoServiceProvider().GetBytes(salt = new byte[16]);
 
             var pbkdf2 = new Rfc2898DeriveBytes(user.Password, salt, 10000);
-            byte[] hash = pbkdf2.GetBytes(20);
+            byte[] hash = pbkdf2.GetBytes(100);
 
             byte[] hashBytes = new byte[36];
             Array.Copy(salt, 0, hashBytes, 0, 16);
